+++ src/Dto/Customer/Wishlist/GetList.php
@@ -37,12 +37,12 @@ class GetList extends Wishlist
      * @var array
      */
     protected $schema = [
-        'type' => 'object',
-        'properties' => [
-            'meta' => ['$ref' => Meta::class],
+        'type'                 => 'object',
+        'properties'           => [
+            'meta'      => ['$ref' => Meta::class, 'skipValidation' => true],
             'wishlists' => [
-                'type' => 'array',
-                'items' => ['$ref' => Get::class]
+                'type'  => 'array',
+                'items' => ['$ref' => Get::class, 'skipValidation' => true]
             ]
         ],
         'additionalProperties' => true